Got all tests working and mypy passing
