+++ guidance/library/_image.py
@@ -1,4 +1,4 @@
-from guidance import guidance
+from .._guidance import guidance
 import urllib
 import typing
 import http
