+++ guidance/library/_optional.py
@@ -1,4 +1,4 @@
-from guidance import guidance
+from .._guidance import guidance
 from .._grammar import select
 
 @guidance(stateless=True)
