+++ guidance/library/_prefix_tree.py
@@ -1,4 +1,4 @@
-from guidance import guidance
+from .._guidance import guidance
 from .._grammar import select, string
 
 @guidance(stateless=True, dedent=False)
