+++ guidance/library/_regex.py
@@ -1,4 +1,4 @@
-from guidance import guidance
+from .._guidance import guidance
 from .._grammar import byte_range, select, string, Byte
 from ._zero_or_more import zero_or_more
 from ._any_char_but import any_char_but
