+++ guidance/library/_role.py
@@ -1,4 +1,4 @@
-from guidance import guidance
+from .._guidance import guidance
 from ._block import block
 from ._set_attribute import set_attribute
