+++ guidance/library/_set_attribute.py
@@ -1,4 +1,4 @@
-from guidance import guidance
+from .._guidance import guidance
 from ._block import block
 
 @guidance
