+++ guidance/library/_set_var.py
@@ -1,4 +1,4 @@
-from guidance import guidance
+from .._guidance import guidance
 from ._block import block
 
 @guidance
