+++ guidance/library/_substring.py
@@ -1,6 +1,6 @@
 from typing import Optional
 
-from guidance import guidance
+from .._guidance import guidance
 # from ._prefix_tree import prefix_tree
 from .._grammar import string, select, capture
 from ._optional import optional
